requirements.txt
absl-py==2.1.0
addict==2.4.0
attrs==23.2.0
basicsr==1.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
facexlib==0.3.0
filelock==3.13.3
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
future==1.0.0
gfpgan==1.3.8
grpcio==1.62.1
humanfriendly==10.0
idna==3.7
imageio==2.34.0
importlib_metadata==7.1.0
imutils==0.5.4
Jinja2==3.1.5
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.42.0
lmdb==1.4.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12>=12.1.3.1
nvidia-cuda-nvrtc-cu12>=12.1.105
nvidia-cudnn-cu12>=8.9.2.26
onnxruntime==1.17.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.25.3
PyMatting==1.1.12
pyparsing==3.1.2
PyQt6==6.6.1
PyQt6-Qt6==6.6.3
PyQt6-sip==13.6.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
realesrgan==0.3.0
referencing==0.34.0
rembg==2.0.56
requests==2.31.0
rpds-py==0.18.0
scikit-image==0.22.0
scipy==1.13.0
six==1.16.0
sympy==1.12
tb-nightly==2.17.0a20240406
tensorboard-data-server==0.7.2
tifffile==2024.2.12
tomli==2.0.1
torch>=2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.3
typing_extensions==4.11.0
urllib3==2.2.2
Werkzeug==3.0.6
yapf==0.40.2
zipp==3.19.1
boto3
pydantic_settings
fastapi==0.111.0
uvicorn==0.29.0
gunicorn==22.0.0
lightgbm==4.5.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
opencv-contrib-python==4.10.0.84
sentry-sdk[fastapi]==2.16.0
xgboost==2.1.1
google-cloud-firestore==2.19.0
psutil==6.0.0